% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_modules.R
\name{compare_modules}
\alias{compare_modules}
\title{Compare WGCNA module builds}
\usage{
compare_modules(
  fit,
  Rsq.min = NULL,
  sft.value = NULL,
  minModuleSize = 20,
  maxBlockSize = 500,
  deepSplit = 3,
  networkType = "signed",
  TOMType = "signed",
  nThread = 2
)
}
\arguments{
\item{fit}{List object output by fit_modules( )}

\item{Rsq.min}{Numeric vector minimum R-squared for soft threshold selection. If set, sft.value is ignored}

\item{sft.value}{Integer vector soft threshold. Set when minimum R-squared is ignored}

\item{minModuleSize}{Numeric vector minimum module size. Default is 20}

\item{maxBlockSize}{Integer vector giving maximum block size for module detection. Default is 500}

\item{deepSplit}{Integer vector between 0 and 4. Provides a simplified control over how sensitive module detection should be to module splitting, with 0 least and 4 most sensitive}

\item{networkType}{Character vector. Network type from "unsigned", "signed", "signed hybrid". Default is "signed"}

\item{TOMType}{Character vector from "none", "unsigned", "signed", "signed Nowick", "unsigned 2", "signed 2", or "signed Nowick 2". If "none", adjacency will be used for clustering. See TOMsimilarityFromExpr for details. Default is "signed"}

\item{nThread}{Integer for number of threads to use}
}
\value{
List including:
\itemize{
  \item{summary} Data frame with total modules and sizes for all builds
  \item{modules} List of data frames with module sizes for each build
}#'
}
\description{
Make WGCNA modules from gene expression data with dynamic soft threshold selection. Also outputs mean module expression and DAVID formatted gene lists
}
\examples{
fit <- fit_modules(dat = example.voom)
compare.mods <- compare_modules(fit = fit, sft.value = c(14, 20))
}
