% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rare.R
\name{filter_rare}
\alias{filter_rare}
\title{Filter rare and low abundance genes}
\usage{
filter_rare(
  dat,
  min_CPM,
  gene_var = "ensembl_gene_id",
  min_sample = NULL,
  min_pct = NULL,
  plot = FALSE,
  min.CPM = NULL,
  gene.var = NULL,
  min.sample = NULL,
  min.pct = NULL
)
}
\arguments{
\item{dat}{DGEList output by edgeR::DEGList( )}

\item{min_CPM}{numeric minimum counts per million (CPM)}

\item{gene_var}{character name for column with gene names in dat$genes that matches names in expression data dat$E. Default "ensembl_gene_id"}

\item{min_sample}{numeric minimum number of samples}

\item{min_pct}{numeric minimum percent of samples (0-100)}

\item{plot}{logical if should plot mean variance trends}

\item{min.CPM}{Deprecated form of min_CPM}

\item{gene.var}{Deprecated form of gene_var}

\item{min.sample}{Deprecated form of min_sample}

\item{min.pct}{Deprecated form of min_pct}
}
\value{
DGEList object filtered to not rare genes
}
\description{
Filter genes at a minimum counts per million (CPM) in a minmum number or percent of total samples.
}
\examples{
dat.filter <- filter_rare(dat = example.dat, min_CPM = 0.1, min_sample = 3,
                          gene_var="geneName")
dat.filter <- filter_rare(dat = example.dat, min_CPM = 0.1, min_pct = 10,
                          plot = TRUE, gene_var="geneName")
}
