% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_modules.R
\name{fit_modules}
\alias{fit_modules}
\title{Perform soft thresholding in preparation for WGCNA module building}
\usage{
fit_modules(
  dat,
  genes = NULL,
  powerVector = c(1:30),
  networkType = "signed",
  nThread = 2
)
}
\arguments{
\item{dat}{limma EList output by voom( ). Must include dat$E}

\item{genes}{Character vector of genes to used in module building. Must match rownames in dat. If not set, all genes in dat are used}

\item{powerVector}{Numeric vector of soft thresholding powers for which the scale free topology fit indices are to be calculated. Default c(1:30)}

\item{networkType}{Character string of network type. Allowed values are "unsigned", "signed", "signed hybrid"}

\item{nThread}{Integer for number of threads to use}
}
\value{
List including:
\itemize{
  \item{genes} Character vector of genes used in module building
  \item{sft} Data frame with soft thresholding results including R-squared, slope, and k metrics
  \item{top.plot} ggplot object of soft thresholding topology
  \item{connect.plot} ggplot object of soft thresholding connectivity
}
}
\description{
Perform soft thresholding in preparation for WGCNA module building
}
\examples{
fit <- fit_modules(dat = example.voom)
}
