% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_module_dat.R
\name{make_module_dat}
\alias{make_module_dat}
\title{Create dat object with mean module/gene set expression}
\usage{
make_module_dat(
  module_df = NULL,
  dat = NULL,
  genecol = "ensembl_gene",
  modcol = "gs_name",
  lib = "libID"
)
}
\arguments{
\item{module_df}{Data frame with module names in one column and gene membership in another column}

\item{dat}{limma EList object to alter}

\item{genecol}{String showing name of the column in module_df containing the gene IDs. Be sure these IDs match the row names in the input dat object. (Default "ensembl_gene")}

\item{modcol}{String showing name of the column in module_df containing the module IDs. (Default "gs_name")}

\item{lib}{String showing the name of the column containing sample IDs in dat$targets. This will become the column names for the new collapsed expression matrix. (Default "libID")}
}
\value{
limma EList object
}
\description{
Create dat object with mean module/gene set expression
}
\examples{
dat.voom <- RNAetc::example.voom
mod_df_ex <- data.frame("gs_name" = c(rep("M1", 5),
                                      rep("M2", 5),
                                      rep("M3", 5)),
                        "ensembl_gene" = c(rownames(dat.voom$E)[1:15]))

make_module_dat(module_df = mod_df_ex, dat = dat.voom)
}
