% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_modules.R
\name{make_modules}
\alias{make_modules}
\title{Construct WGCNA modules and associated data}
\usage{
make_modules(
  fit,
  Rsq_min = NULL,
  sft_value = NULL,
  minModuleSize = 20,
  maxBlockSize = 500,
  deepSplit = 3,
  networkType = "signed",
  TOMType = "signed",
  mods_mean = FALSE,
  mods_eigen = FALSE,
  david = FALSE,
  nThread = 2,
  Rsq.min = NULL,
  sft.value = NULL,
  mods.mean = FALSE,
  mods.eigen = FALSE
)
}
\arguments{
\item{fit}{List object output by fit_modules( )}

\item{Rsq_min}{Numeric minimum R-squared for soft threshold selection. If set, sft_value is ignored}

\item{sft_value}{Numeric soft threshold. Set when minimum R-squared is ignored}

\item{minModuleSize}{Numeric minimum module size. Default is 20}

\item{maxBlockSize}{Integer giving maximum block size for module detection. Default is 500}

\item{deepSplit}{Integer value between 0 and 4. Provides a simplified control over how sensitive module detection should be to module splitting, with 0 least and 4 most sensitive}

\item{networkType}{Character string. Network type. One of "unsigned", "signed", "signed hybrid". Default is "signed"}

\item{TOMType}{Character string. One of "none", "unsigned", "signed", "signed Nowick", "unsigned 2", "signed 2", or "signed Nowick 2". If "none", adjacency will be used for clustering. See TOMsimilarityFromExpr for details. Default is "signed"}

\item{mods_mean}{Logical if include mean module expression in output}

\item{mods_eigen}{Logical if include eigenvalue module expression in output}

\item{david}{Logical if include DAVID formatted genes in modules in output}

\item{nThread}{Integer for number of threads to use}

\item{Rsq.min}{Deprecated form of Rsq_min}

\item{sft.value}{Deprecated form of sft_value}

\item{mods.mean}{Deprecated form of mods_mean}

\item{mods.eigen}{Deprecated form of mods_eigen

\itemize{
  \item{genes} Character vector of genes used in module building
  \item{sft} Data frame with soft thresholding selected for module building. Includes power, minimum R-squared, and connectivity
  \item{top.plot} ggplot object of soft thresholding topology
  \item{connect.plot} ggplot object of soft thresholding connectivity
  \item{mods} Data frame of genes in modules
  \item{mods.mean} Data frame of mean module expression for each library
  \item{mods.eigen} Data frame of module eigenvalue expression for each library
  \item{david} DAVID formatted data frame of genes in modules
}}
}
\description{
Make WGCNA modules from gene expression data. Also outputs mean or eigenvalue module expression and DAVID formatted gene lists
}
\examples{
fit <- fit_modules(dat = example.voom)
dat.mods <- make_modules(fit = fit, sft_value = 14,
    mods_mean = TRUE, mods_eigen = TRUE, david = TRUE)
}
