% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fcounts.R
\name{read_fcounts}
\alias{read_fcounts}
\title{Read in counts table from Subread featureCounts}
\usage{
read_fcounts(file)
}
\arguments{
\item{file}{Character string for file path of counts table}
}
\value{
data frame
}
\description{
Read in counts table from Subread featureCounts
}
