% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_voom.R
\name{subset_voom}
\alias{subset_voom}
\title{Subset voom object}
\usage{
subset_voom(
  dat_voom,
  lib_keep = NULL,
  lib_remove = NULL,
  lib_filter = NULL,
  gene_keep = NULL,
  libraryID = "libID"
)
}
\arguments{
\item{dat_voom}{Voom object to subset}

\item{lib_keep}{vector of library IDs to keep (Default NULL)}

\item{lib_remove}{vector of library IDs to remove (Default NULL)}

\item{lib_filter}{string to use for filtering libraries to keep (Default NULL)}

\item{gene_keep}{vector of genes to keep (Default NULL)}

\item{libraryID}{name of the column with library IDs (Default "libID")}
}
\value{
voom object
}
\description{
Subset voom object
}
\examples{
dat.voom <- kimma::example.voom

subset_voom(dat.voom,lib_keep = c("lib1","lib2"))
subset_voom(dat.voom,lib_remove = c("lib1","lib2"))
subset_voom(dat.voom,lib_filter = "asthma == 'healthy' & virus == 'none'")
subset_voom(dat.voom,gene_keep = c("ENSG00000000460", "ENSG00000001460"))
subset_voom(dat.voom,lib_keep = c("lib1","lib2"),
  gene_keep = c("ENSG00000000460", "ENSG00000001460"))
}
