% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_module_coherence.R
\name{calculate_module_coherence}
\alias{calculate_module_coherence}
\title{Calculate WGCNA module coherence}
\usage{
calculate_module_coherence(
  module_gene_sets,
  voom_obj,
  gene_set_column = "mod.color",
  ensembl_ID_column = "geneName",
  module_set = "STUDY",
  sample_set = "STUDY",
  remove_sets = NULL,
  return_plot = TRUE,
  r_cutoff = 0.3,
  p_cutoff = 0.01
)
}
\arguments{
\item{module_gene_sets}{Dataframe giving module membership. Should contain columns specifying ensembl gene IDs & module definitions.}

\item{voom_obj}{Voom object containing expression dataset in which module coherence is to be tested. Should either me a voom EList with "E" expression matrix, or an expression matrix.}

\item{gene_set_column}{Character string specifying name of the column in module_gene_sets df which defines modules. Defaults to "mod.color" which is output by make_modules()}

\item{ensembl_ID_column}{Character string specifying name of the column in module_gene_sets df which gives gene IDs. Defaults to "geneName" which is output by make_modules()}

\item{module_set}{Character string specifying name the study/dataset from which modules were built. This is used simply for labeling of outputs. Default = "STUDY"}

\item{sample_set}{Character string specifying name of the study from which the data come. This is used simply for labeling of outputs. Default = "STUDY"}

\item{remove_sets}{Vector of character strings naming modules which you want removed from the analysis, eg. "0" or "Grey".}

\item{return_plot}{Logical indicating whether plot should be printed when function runs. Defaults = TRUE.}

\item{r_cutoff}{Vector or single numeric value at which to draw red cutoff lines in correlation plot. Default = 0.3}

\item{p_cutoff}{Vector or single numeric value at which to draw red cutoff lines in significance plot. Default = 0.01}
}
\value{
List including:
\itemize{
  \item{coherence_boxplot_combined} A ggplot plot object visualizing distributions of correlation and p values.
  \item{coherence_boxplot_cor} A ggplot plot object visualizing distributions of correlation values.
  \item{coherence_boxplot_p} A ggplot plot object visualizing distributions of correlation p values.
  \item{subgene_correlation_df} Dataframe containing pairwise gene correlations within modules.
  \item{GSabsent} A list of missing genes enumerating genes defined in modules but missing from expression data.
}#'
}
\description{
Calculate within module pairwise gene correlations from WGCNA modules and gene expression data.
Test the coherence of modules built from Dataset A in data from Dataset B.
}
\examples{
moduleCoherence <- calculate_module_coherence(module_gene_sets = example.mods,
voom_obj = example.voom, module_set = "Example WGCNA Modules",
sample_set = "Example Data", r_cutoff = 0.3, p_cutoff = 0.01)
}
