% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_voom.R
\name{combine_voom}
\alias{combine_voom}
\title{Combine data frame elements of a voom object}
\usage{
combine_voom(dat, libraryID = "libID", geneID = "geneName")
}
\arguments{
\item{dat}{EList voom object resulting from voom() or voomWithQualityWeights()}

\item{libraryID}{Character. Column name in dat$targets that matches column names in dat$E}

\item{geneID}{Character. Column name in dat$genes that matches row names in dat$E}
}
\value{
data frame
}
\description{
Combine data frame elements of a voom object
}
\examples{
combined <- combine_voom(example.voom)
}
